% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-feed.R
\name{ino_add_feed}
\alias{ino_add_feed}
\title{Subscribe to feeds}
\usage{
ino_add_feed(feed)
}
\arguments{
\item{feed}{to subscribe to (e.g. \verb{feed/http://feeds.arstechnica.com/arstechnica/science})}
}
\value{
(invisibly) subscription request result as a classed (\code{ino_sub_res}) list
}
\description{
Subscribe to feeds
}
\references{
\href{https://www.inoreader.com/developers/add-subscription}{Add subscription}
}
