% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ino_auth}
\alias{ino_auth}
\title{Authenticate to Inoreader's API}
\usage{
ino_auth()
}
\value{
the OAuth token (invisibly). NOTE: You do not need to use this; tis provided merely
as a convenience
}
\description{
This uses Inoreader's \href{https://www.inoreader.com/developers/oauth}{OAuth 2.0 endpoint} and
MUST be called at the start of a session before any other operations.
}
