% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-subscription.R
\name{ino_edit_subscription}
\alias{ino_edit_subscription}
\title{Edit subscription}
\usage{
ino_edit_subscription(
  action = c("edit", "subscribe", "unsubscribe"),
  stream_id,
  title = NULL,
  add_to = NULL,
  remove_from = NULL
)
}
\arguments{
\item{action}{Action. Can be \code{edit}, \code{subscribe}, or \code{unsubscribe}.}

\item{stream_id}{Stream ID, e.g. \verb{feed/http://feeds.arstechnica.com/arstechnica/science}}

\item{title}{Subscription title. Keep \code{NULL} to keep the title unchanged}

\item{add_to}{Add subscription to folder. Use full folder name like \verb{user/-/label/Tech}.
Keep \code{NULL} to leave in place.}

\item{remove_from}{\itemize{
\item Remove subscription from folder. Use full folder name like \verb{user/-/label/Tech}
Keep \code{NULL} to leave in place.
}}
}
\value{
(invisibly) subscription edit request result as a classed (\code{ino_edit_sub_res}) list
}
\description{
This method is used to rename the subscription, add it to a folder, remove it from folder or
unsubscribe from it. You can also subscribe and directly rename it and add it to folder via
this method.
}
\references{
\href{https://www.inoreader.com/developers/edit-subscription}{Edit subscription}
}
