% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder-tag-list.R
\name{ino_folder_tag_list}
\alias{ino_folder_tag_list}
\title{Fetches the current folders and tags for the logged-in user}
\usage{
ino_folder_tag_list(types = TRUE, counts = TRUE)
}
\arguments{
\item{types}{(lgl) if \code{TRUE} (the default) the item type will be included in the result}

\item{counts}{(lgl) if \code{TRUE} (the default) unread counts for tags and active searches will be included
in the result}
}
\value{
data frame. See linked reference for field definitions.
}
\description{
Fetches the current folders and tags for the logged-in user
}
\references{
\href{https://www.inoreader.com/developers/tag-list}{Folders and tags list}
}
