% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-contents.R
\name{ino_stream_contents}
\alias{ino_stream_contents}
\title{Returns the articles for a given collection}
\usage{
ino_stream_contents(
  stream_id,
  num_items = 20L,
  order = c("newest", "oldest"),
  start_time = (Sys.time() - 604800),
  exclude = NULL,
  include = NULL,
  continuation = NULL,
  all = TRUE,
  annotations = FALSE
)
}
\arguments{
\item{num_items}{Number of items to return (default \code{20}, max \code{1000}). If the value passed is
less than zero or greater than one thousand, 20 and 1000 will be used (respectively).}

\item{order}{Order. By default, it is \code{newest} first. You can pass \code{oldest} here to get oldest first.}

\item{start_time}{Start time (\code{POSIXct} or a string that can be converted to \code{POSIXct}) from
which to start to get items. Defaults to 7 days ago.}

\item{exclude}{Exclude Target — You can query all items from a feed that are not flagged as read
by setting this to  \verb{user/-/state/com.google/read}. Default is \code{NULL}.}

\item{include}{Include Target — You can query for a certain label with this. Accepted values: (e.g.)
\verb{user/-/state/com.google/starred}, \verb{user/-/state/com.google/like}. Default is \code{NULL}.}

\item{continuation}{Continuation (default \code{NULL}) — a string used for continuation process.
Each response returns not all, but only a certain number of items.
Pass the \code{continuation} value of the returned list (if one exists) to retrieve the
next part of the result set. If there is no \code{continuation} value, there are no more
feed items left.}

\item{all}{\itemize{
\item set this to \code{FALSE} (default is \code{TRUE}) if you want to receive only manually added
tags in the categories list. Otherwise automatically added tags from the folders will be
populated there too.
}}

\item{annotations}{\itemize{
\item set this to \code{TRUE} (default is \code{FALSE}) if you want to get an array of your
annotations for each article.
}}
}
\value{
list with metadata about the result and result items in the \code{item} slot. See linked
reference for field information.
}
\description{
Returns the articles for a given collection
}
\references{
\href{https://www.inoreader.com/developers/stream-contents}{Stream contents}
}
