% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unread-counts.R
\name{ino_unread_counts}
\alias{ino_unread_counts}
\title{Retrieve unread counters for folders, tags and feeds.}
\usage{
ino_unread_counts()
}
\value{
data frame. See linked reference for field definitions and the meaning of the \code{max} column.
}
\description{
Retrieve unread counters for folders, tags and feeds.
}
\references{
\href{https://www.inoreader.com/developers/unread-counts}{Unread counters}
}
