% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-info.R
\name{ino_user}
\alias{ino_user}
\title{Retrieve basic information about the logged in user}
\usage{
ino_user()
}
\value{
(invisibly) user information as a classed (\code{ino_user}) list
}
\description{
Retrieve basic information about the logged in user
}
\references{
\href{https://www.inoreader.com/developers/user-info}{User information}
}
